requirements.txt
pipenv
flask
python-dotenv
sqlalchemy
flask-sqlalchemy
wtforms
flask-wtf
flask-login
werkzeug==2.3
flask-migrate